The trickiest bit of all was to get the cancel edit on blur working. It was tricky because of the multiple ajax request timing problem. Two tricky cases to handle: 1. hitting enter shouldn't do the cancel edit because if the cancel edit ajax returns before the edit ajax then it would look like the edit didn't happen 2. clicking another todo for example shouldn't cancel edit because if the cancel edit ajax returns first it will change the url and then the edit ajax return won't be applied because the url's been changed
For the first, the edit request fires first so added an attribute to the input box that's checked in the blur.
For the second, the cancel edit fires first so set a boolean that can be unset if a subsequent ajax is made before the cancel edit ajax returns. If so then don't apply the cancel edit response because the subsequent ajax will cancel the edit anyway. Tweaked the navigation js so ajax resp can be cancelled by setting panels to null.

Another tricky part was that couldn't keep the other RefreshPanels always refreshing because if clicking out of the edit into one of these RefreshPanels then the setting of resp of the cancel edit ajax will get in the way of the click. For example, couln't click the completed hyperlink if this was refreshed. So, adding a Func to prevent these panels being refreshed if clearing - only the edit todo panel needs refreshing.