Only add history if urls match
Might be that navigate within a navigated. In that case the first
navigate history is superseded and shouldn't be added. Bug showed up
when page first loaded and the HashHistoryManager got in a muddle
thinking it was HTML5 url because the urls didn't match.
Didn't really want to navigate within navigated so moved name
subscription until after navigated fired.