Required 'angular' and 'jquery' and shimmed angular
Without requiring jquery the build failed when compiling
IAugmentedJQuery on LinkUtility. Tried using the angular npm package -
as done with knockout an react - but the angular module failed.  Tried
just require('angular') independent of package and that failed! It tries
to access the non-existent window, sure the angular npm project's broken