Fixed compile/gulp task problem
Couldn't get it so the compile worked and the gulp tasks. Could get one
or the other but not both. Traced it back to the reference path to
navigation.d.ts in NavigationPlugins. It messed up the required of
navigation in test.ts files. Without it the package command wouldn't
run, so manually added the file in in the gulp task