Reset history if null url is passed to addHistory
To cancel a history navigation in unloading/navigating need to rollback
the url because pressing the back button changes the url straight away.
Calling addHistory(null, null) rolls it 'back' by setting StateContext
url as current url.
Didn't set the State if null passed in because the document title
might've been changed and don't want to reset it.