Changed parse logic so all path parts are available
Need to removed the regex from the Segment build so need non-param and
param parts of the path available. Bit messy at the moment bit it's
matching all parts of path instead of just replacing params.
Tweaked reserved char unit test because it shouldn't have had curlies in
it, these are used to specify params