Cached route parse result
Tracked the group of params that make up all the routes for a State.
Then intersection these with the incoming data to detect the params key
for the cache.
route1 = '{sort}'
route2 = '{sort}/page/{page}'
then the cached params would be 'sort' and 'page'. Incoming data might
be { sort: 'name', filter: 'bob' }, then the cache key would be 'sort'.
No need to include filter as part of the key