Moved more route tests to use public Navigation api
The reserved url ones weren't routine like the others because:
1. comma are reserved for route splitting so removed
2. underscores in values get replaced with 0_ so they don't conflict
with separators for preserving data type
Need to look back at 2 and see if it's the right thing to do. People
won't want their underscores being changed. Now there's only strings,
booleans and numbers, and seeting that it's only the strings that can
have underscores, there's no need to convert them if the default type is
string? something to think about