Changed State route to union type
Can define multiple routes as an array rather than as a comma delimited
string. This changed showed a bug with the sorting of routes that puts
the 'catch-all' {x} at the bottom. It needed to account for all the
routes not the routes string. Could specify the catch-all route as the
second one. ['abc/{x}', '{x}'].
The trick was to sort the routes on the Router after they're build
rather than trying to sort them on the States before they're built.