Fixed two Route track types picking wrong route
The default value check in the CrumbTrail and ReturnData managers was
checking the values before formatting. A default value of 2 and new
value of '2' aren't === so it will be passed to the router and the wrong
route is picked.
With track types true these aren't the same value. But with track types
false these two are the same value, because they'll both appear in the
Url the same and get converted using the default type at the other side.
Changed the default check to run after the formatting and then check
against the formatted default instead of the default.