Added changes into NavigationLink
The page size links didn't get hrefs to start with when set to lazy.
Wanted to check this was just a refresh navigation thing. After changing
them to NavigationLinks they did get hrefs. A RefreshLink that is on the
screen before a navigation won't get an href if set to lazy because it
won't run after a navigation - should fix this? Probably not because
even if run an initial navigation there's no guarantee that the
RefreshLink will return non null