Added listeners conditionally based on lazy flag
If the onNavigate listener's always added and the lazy flag checked
inside then when there's a lot of links (10s of 1000s) then looping
through that many listeners and checking the flag still adds many
milliseconds (5 ms for 10,000). The idea of lazy is it won't impact
navigation at all. Changed so that listeners aren't added at all if lazy
flag set. The only downside is that lazy can't be an observable - can't
dynamically change it - but that's not a likely scenario