Improved react sample code
Not part of lazy link, but had to try out masterpage idea for react.
Moved the navigated listeners outside of components and created
components in them. Attached them both to the same dom node. That's the
masterpage idea - define the html masterpage template with placeholders
and set components into the placeholders in the navigated handlers.
Components are clean, no routing code, no state, just props passed in