Checked left click without modifiers
Read that the left click event button isn't 0 in ie < 9
(http://www.quirksmode.org/js/events_properties.html#button).
Tested hyperlink click in ie6 - 8 (using browserstack) and the left
click is 0; and the middle click doesn't even fire the onclick event.
The confusion is around when clicking div, not a link or a button,
that's when the event button isn't 0 in ie < 9 - ran the test page,
http://www.w3schools.com/jsref/tryit.asp?filename=tryjsref_event_mouse_button,
and the event buttons are 1,4,2 in ie < 9.
To summarise, it's safe to check for e.button being 0 because that means
left click (or enter key) in all browsers