Always forced lazy update
Not technically necessary, but prefer to force an update when the href
changes. Means that the link's in a consistent state - not just the href
updated but related updates also run.
Noticed it when activeCssClass and lazy both on - only href updated, not
the class name. In reality, can't set active properties if lazy's on.
But forcing update feels right