Passed previous data in back and refresh links
The previous data changed when the combineCrumbTrail setting changed.
Here's what the previous data and old data represent
- previous data holds the context data at the time the link was built
- old Data holds the context data at the time the link is executed
The previous and old data should match, provided not using a bookmarked
link. So, the combineCrumbTrail setting shouldn't effect them. When it
was set to true then the crumb trail was built to include the current
data and added to the url. When set to false the crumb trail links
didn't pass the current data into GetHref so wasn't part of Url. Changed
the GetHref so that the return data isn't passed in, the current
StateContext data's always used.
It's kind of a breaking change because the generated Urls are different.
But the old Urls still work so not sure?