Accepted HistoryAction as a string
When putting the changes into Angular couldn't set the typescript enum
directly as the attribute value, the context is the scope and can't be
the window. Don't want to force users to create a history action
property on the scope just to bind. One option was to pass the number in
(enums are numbers) but that would be hard to understand. So, angular
needed to accept strings, but don't want the plugins to have different
surfaces so React and Knockout also need to accept strings