Built array for multiple query string values
This found an unrelated bug. The Router added in the default values but
the State Handler didn't. That meant defaults for route parameters
handled differently to query parameters. For example, a default of '2'
with default type of 'number' would be turned into 2 if it's a route
parameter but left as '2' if query string. Moved default setting into
State Handler so it's handled consistently, but need to revisit and add
tests