Removed default setting from State Handler
Already removed from Router. This means all default setting done by
StateController calling setDefaults. Problem with that is the router and
handler can still set defaults if it appears in the route or query
string - either manually, by changing the url or the router can put
default values when it has to pad the url. So changed the
StateController to not pass default values into the parse method - that
way defaults always ignored on url. Means the default value is never
parsed and always taken from the defaults - so can have default 'a' and
default type 'number' and not throw errors. Also, default '2' and
default type 'number' should have '2' as default not 2. Updated the 4
failing tests to reflect that