Encountered array size 1 query string problem
An array of size 1 appears differently if it's a route param or query
string. If it's a route param then - is replaced with 2- to cater for
separator of 1-. Trouble is by the time it reaches the ArrayConverter
can't tell whether it came from route or query string so don't know
whether to convert 2- back to - or not.
Haven't added a test for this yet. So just added code to identify query
string params but won't use it until there's a failing test in place