Added urlDecode function to Router
Passed it in as optional parameter and then passed down to Route.  The
getData of StateRouter doesn't know the State and isn't always called
from the StateHandler - navigateLink calls it to get the State. However,
can get the State from the route and can then call urlDecode on
StateHandler. Had to add urlDecode to IStateHandler - and made it
optional for backward compatibility