Added Navigation hot loading
Reloads the new StateInfo and State Navigators. Then the cool part - it
replaces the 'old' states and dialogs on the StateContext with the 'new'
ones. It uses the keys from the 'old' to find the corresponding 'new'.
So, if the StateContext state was masterDetails-listing then it finds
the masterDetails-listing state in the new config and replaces it on the
StateContext. Once all the 6 StateContext items are replaced it does a
refresh navigation using the current StateContext data. This updates the
Url and triggers the updated State Navigators