Only updated State Info if error free
Inspired by hot loading. If there's an error in State Info then don't
want to throw an error in HMR (webpack hot module replacement) because
it refreshes the page and won't help fixing the State Info error. Better
to catch and log the error but not throw an exception and not continue
with the hot replacement, i.e., don't update the StateContext. That way,
the old State Info remains and the app works as before, allowing State
Info error to be fixed and the next hot load to apply it