Catch hot loaded StateInfo error and stop
If the StateInfo is in error then don't want to throw the error because
that reloads page and still leaves the StateInfo error. So catch and log
error but don't reload the Context. This won't work until the StateInfo
build change comes in because the error StateInfo overwrites the old