Added arrayData to StateRouter
Similar to the queryStringData added to StateHandler to support
splitting array on query string. This is to support splitting array on
route for a splat parameter. Don't need to pass data and arrayData to
router separately because the router doesn't need both - if it's a splat
it wants the arrayData otherwise the data