Built combined data per Route - not per State
A State can have multiple routes and the same parameter can be a splat
in one of them and not a splat in the other. Need to pass the array data
to the splat route but the non-array to the non-splat. Also applies to
matching, need to work out separable data based on the Route, not the
State. So, used Route instead of RouteInfo