Changed Default value check
Example: single item string array whose item value matches the formatted
value of the default two item array value. The default check thought the
value passed was default so didn't parse it and the default two item was
set in navigation data. Changed the check to assume separable (not
combineArray) is never the default value. This would only go wrong if
set a default array value with a conflicting default type because it
would try to parse the array with the default type - obscure, but should
check