Default check shouldn't match string and array
Single item array in splat is a string and default array value is always
an array.  So string 'ab' != ['a','b'] - but the isDefault check made
them the same because it checked lengths and each item. Added a typeof
check to isDefault. The formattedArrayDefault is always an array even if
it just has one item so no need to check its type