Added stop to history manager
Now there can be more than one state controller, need to stop a history
manager. Also, can change history manager even within a single state
controller, so that means need to stop the old and start the new. Don't
want the plugins to pass the application path in because would have to
make settings public, so passed in as part of the init