Changed meaning of Previous State!
Used to be that previous state == old state (except when using
bookmarked link). Now, back navigation uses links from url - it doesn't
rebuild them - so previous state is what is was not what it is! Not only
is this better but it allows for repeated states in the crumb trail.
Couldn't do this before because back navigation was really forward
navigation to a repeated state so had to truncate. Now, back navigation
is just like a bookmarked link