Created crumbs with navigation data
Could have two states with same route 'a' with crumb trail turned on on
first route. Might pass { crumb: 'hello'} to the second state where the
crumb key is used but it's not a crumb trail. Then parsing Url of
'/a?crumb=hello' would cause problem because it would match first route
but wouldn't be a valid crumb trail. So need to move crumb trail parsing
into the navigation data parsing so can submit it to validation and
continue matching if validation fails.