Propagated error message if match fails
Need to know why the match failed, not just that the Url is invalid. At
the moment only showing the last error message - there might be more
than one now that constraints are around because the match continues
after a match and error.
Removed the global error handler in History Manager. Better to wrap
start in a try catch and replace the Url. That should be only way to get
an error. The global error handler is too crude because no idea where it
came from. Probably add error handlers to the plugins so can handle at
the Link level because can then act on the error