Set state based on whether context url changed
Previously the Links were rendering twice. Once when they render and
then again because the onNavigation listener fired. This only happened
on initial navigation (not refresh) because that's synchronous -
subsequent React renders are async so the onNavigate fires before the
render and React only renders once.
The Back Link still renders twice because it's possible that navigating
the same Url can change the Back Links if there's a custom truncate url
that allows repeated states