Added crumb to onNavigate setState check
The Back Link doesn't need to re-render if the crumb hasn't changed.
But, back link might have active props in future and these could change
even if crumb hasn't (could navigate to same url as crumb but the crumb
trail could still have that crumb in it if there's custom truncation).
Including the stateContext url in the check caters for this and matches
with Refresh and Navigation Links