Changed Animation from HOC to component
When it was a higher order component, a new TransitionMotion class was
created each time and so it never animated? Well, it kept going into
createClass instead of passing in new props.
I guess the problem is that the wrapped People is different to the
wrapped Person. So React deletes instead of updates