Created basic files for web/native sample
Changing Navigation React Native is cumbersome because have to package
it, copied it to sample project then rerun react-native run-iso
command. There’s nothing specific to React Native about it, so creating
a web project to make it easier to test changes. Plus, it’s good to
show this idea works on the web because shows web can do what native
does