Took state and data from styleThatLeft
Take states A, B and C with different leave animations. Navigate from A
to B to C and then got back twice quickly. What happens is the B starts
leaving correctly but when the second back comes in it follows A
leaving trajectory. There’s only one willLeave style so the second
render means B follows A out the door because oldState is now A. The
trick is to use to take the state (and data) from the styleThatLeft
instead of oldState (and oldData). The styleThatLeft has B in it