Passed colorRef to details in prevSceneData
The native handle of the selected color is passed from Grid to Details.
The details before navigating back measures the position of the color
and moves the scene so these figures are stored. Then it navigates back
which uses this scene data to change the target of the unmount style