Used latest scene data to build styles
Although the zoom back was working before, it was relying on
synchronous nature of setState. Pressing X from Details called
moveScene followed by navigateBack. The moveScene ran the render and
started the animation which navigating back then revealed. If setState
was asynchronous then render would only run after both moveScene and
navigateBack updated scenes and so using the sceneData held in
TransitionMotion would break the animation - it would be the old
sceneData and not the one passed in moveScene.
React Fibre makes setState asynchronous so need to cater for it by
using the most up to date scene data when building unmounted style.
Only need TransitionMotion to track the url. The url can’t go state
because it doesn’t change for a given scene. But the sceneData is
intended to change so use the url to get the latest sceneData from state