Worked out when a shared element has animated
If navigating forward then use onRest of the Motion component. But if
navigating back then the old scene might animate away before the shared
element completes. In this case the shared elements will got to 0
because NavigationMotion clears out the old shared elements. Checked if
shared elements are 0 but previous aren’t