Forced shared element motion to call onRest
If a shared element is in the same start and end position then it
didn’t animate and so onRest wasn’t called which stopped the Modal from
being removed. Forced the animation to start by adding a dummy __force
animating value. As long as it’s different to default and uses spring
it works