Switched Motion for Animate component
Tricky bit was setting immutable to false on Animate component. If
don’t set this then every time the SharedElementMotion sets State the
Animate updates which resets the timer progress. So a 300ms animation
takes longer because it resets 300 multiple times. Setting immutable
makes it compare the data props on values instead of ===. Because the
shared elements are going to the same place each update, the data’s the
same so it doesn’t update and so doesn’t reset the timer