Removed move scene
Added this for iOS gesture but it doesn’t work for that because there’s
more to iOS gesture - like title moving along with gesture. Can’t
replicate this so removing move scene because it complicates the code
for no benefit. The only way to replicate iOS gesture is with native
swift code so will think about adding this later