Sorted items by incoming order with leave last
If item A takes a long time to leave then can press back and forward a
few times so that item A doesn’t have time to leave before it comes
back again. Then the items that arrive before A were in front because A
hadn’t left so new items (although preceding A) were rendered later.
Ensure that new and staying items are ordered by the order of the new
props data. The leaving items are put last, with their internal order
copied over from previous render.