Used crumb length as scene index
Can’t use url because every navigation changes the url, adds to scenes
and causes them to animate in and out. Want to use the crumb trail to
determine when scenes animate. So refresh navigation with custom crumb
trail truncation stops the scene leaving.
Had to change motion to use the new data over the old because want the
scene to update for the old key