Stored navigation link in state
The href can be different from the link. On mobile, the href won’t have
the crumbs. So want to navigate the link not the href. Can’t just
re-evaluate it though because this can throw errors - for example,
double click a back link and the second click will throw an error when
re-evaluating because can’t go back anymore. So stored the link in
state so don’t have to re-evaluate (equivalent to pushState changes for
Mobile history manager)