Updated link state only if in current crumb
With Mobile, if you double click link A to go to B then could create A
-> B -> B instead of A -> B. Links should be idempotent, safe to click
multiple times. Achieved this by only updating the link (and active
flags) if the link is on the current crumb. So don’t update link A if
navigated to B. Clicking A again will navigate along the same link and
so will create A -> B.
Same story for back link. Going back from B to A could disable back
link on B as it navigates to unmounted style because can’t navigate
back anymore. Stopping link B from updating when on A prevents this