Smoothed shared transitions across history
On the zoom sample navigate list -> details orange -> list -> details
yellow then press browser back twice and forward twice and repeat.
This goes from orange to yellow but the animation seems to ‘jump’ and
stop short.
The yellow starts animating out and the details starts leaving. When
the details orange is hit the details starts entering from where it
was, e.g., halfway through but the orange starts animating in from 0.
So the details ends when the shared element is only halfway through.
Changed Motion to accept a progress. It assigns the progress to new
elements so that they don’t have to start from 0. Tracked the progress
of the current state in NavigationMotion and passed this through to
SharedElementMotion.
This also works late discovered shared elements - they’ll start from
the progress point of the early discovered shared elements. So all
shared elements will finish together