Added active flag and set overflow from it
Active flag can’t be determined using style or crumbs. With style can’t
tell if it’s coming or going in all scenarios. With crumbs it’s not
safe to get off context because that won’t necessarily match the state
of NavigationMotion. For safety, added an active prop to callback.
Needed it for setting overflow style. Chrome allowed crumb scenes to be
scrolled if overflow auto so stopped by turning off scrolling when
inactive