Got medley translate working on chrome
Using emulator showed that the scenes in Medley didn’t take up the
whole screen. Thought this was emulator bug (because it worked on
Firefox emulator) but saw same on BrowserStack. Setting initial-scale
fixed it on emulator and BrowserStack.
Noticed that Medley didn’t work on <= iphone 6s. It expects vendor
prefix. Adding WebkitTransform to style fixed it. Don’t want to add to
the sample because unnecessarily confusing