Smoothed shared els when position changes
Try out the Zoom sample and, when pressing x, scroll the screen so that
the target position of the shared elements change during the animation.
Before change, the shared els didn’t reach their target. That’s because
when their targets change then their progress was reset.
Took the progress from the prop passed in because that represents the
progress of the navigation animation. This half-worked. The problem was
that the animation jumped to its next position. It was taking it’s new
start position from the item’s style. This works for the navigation
animation, when animating from mount to crumb. But it doesn’t work here
where need to keep the same start style.