Kept with onNavigate listener
Tried changing to componentWillReceiveProps and it did work, but it was called multiple times. Not sure about this. Anyway, componentWillReceiveProps is deprecated for getDerivedStateFromProps but can't change to this because won't work in old React (and it's called multiple times). When come to do SSR then need it to work with a navigation, first render, so can revisit then