Moved raf from willReceive to didUpdate
I had so much trouble with this becuase of a bug with the new context api. With this change the raf runs forever, even after the animation completes. Thought that meant couldn't use this approach but really just a bug with context api. See https://github.com/facebook/react/issues/12335. Need to retest this when next alpha comes out and check the raf stops running