Attempted immutable navigation
Overrides navigateLink on stateNavigator and calls setState with new StateContext. Creates a new stateNavigator cloned from original and sets on new StateContext. Overrides this new navigateLink and so it goes on.
Don't finish the navigation - add history, call navigated etc - until the setState completes.
There's a problem with the history callback firing because it's pointing at the original stateContext