Changed shared element progress 1 to 0
When navigating back from zoom details to list, the shared element motion tick now fires before the navigation motion tick. That's because the shared el tick is registered in componentDidMount but the navigation tick is registered in componentDidUpdate (they used to both be registered in componentWillMount). So when the shared el tick fires the progress is still set to 1 but it should be 0