Deferred set state
Used the experimental unstable_deferredUpdates. If defer is true then call an immediate setState (sync) with the current state context and the next state and data so the current view can be updated with spinner. And defer a set state for the new view with the new context so React can start rendering in background