Added odl url changed check to setState
Click deferred link and the immediate setState is called once. But click deffered link twice and it's called three times! I think it's because the second one comes after the first deferred, so when it comes to doing the second deferred it wants to apply it on top of the latest state so it calls the second immediate setState again and then applies the second deferred on the top. But the first deferred has already run (not sure if this is the real behaviour) so when it runs the second deferred the previousState is already updated so the third call is with the 'wrong' state.
Don't think it matters because then the deferred is applied which doesn't use the previousState and so overwrites the 'wrong' values. So can't think of a unit test that would break. But still want to apply this check so that the third time it runs it returns null and the state isn't changed. It mirrors the url checks in the navigateLink of StateNavigator.