Derived next items from props
Need the first render to be sync when props change so that the new Scene is given the right NavigaitonContext.
If it's not synchronous then can do these two navigations in a row:
navigateLink(B?crumb=A)
navigateLink(C?crumb=A&crumb=B)
The first navigateLink sets B to render but, before it can, the second navigateLink comes in. When it comes to render then it will render B and C together for the first time and so B will get the wong NavigationContext.