Stored navigationEvent in parent state
Instead of Scene storing navigationEvent in its state, stored it in NavigationMotion state. This means that don't have to wait until the render and Scene is constructed to create the NavigationContext value. As soon as the navigation happens create the NavigationContext value, store it and then it doesn't matter when the render happens and the scenes created. Not sure if it was a bug before, but a lot can happen between getDerivedStateFromProps and render - maybe another navigation could happen from a child component that could trigger before the render? Anyway, this guarantees each scene gets the right context. Also, not limited to rendering one scene at a time - it's a good restriction for now, but might be cases where want to render more than one