Decoded '' for blank value
Don't want to pass null (or undefined) to decode because decodeURIComponent(null) is 'null'. If someone forgot to put a matching decode for the blank encode then would end up with a value of 'null' and confusion. With '' you'd realise you hadn't decoded the blank.
Plus a query stirng of ?a is handled the same as ?a= Can't generate the second one but it could be passed in