Pointed at latest react + navigation
If type in route as 'xx{like' then the toggle link should disable because that's an invalid link. But the new consumer and immutability meant that the NavigationLink didn't get the updates stateNavigator so it still worked off the last valid route. Ensure it updated by passing in a different with each render to clear out the state