Navigated back multiple crumbs
The only way could find is NavUtils.navigateUpTo and pass the target Intent. Means have to keep track of the intents as they're added (and removed).
Also, navigateUpTo closed the app when going back to first Intent?! Don't know why. Changed the launchMode of MainActivity to singleTask and it fixed it