Got navigateUpTo to reuse activity
If launchMode is standard (default) then the up activity is popped off and a new one created. So changed launchMode to singleTop. Problem is with singleTop can't use the same Scene activity class twice in a row because it won't add to stack. So created Odd and Even scene activities to alternate