Allowed reloading
When reload the javascript is reinitialised but the native isn't. So the scene in native might be 2 but in javascript it's 0 so it renders blank. Changed handle back for this scenario so it's handled in native and can then navigate back to first scene and continue from there.
No point doing anything more complex because the javascript's been initialised so should start from first scene. Developer can change first navigation to start on different scene