Changed to did show
The problem is that willShow fires when peeking back using swipe. If you peek back and then stop swiping so you stay where you are then the native is out of step with the js because js navigates back when native doesn't. Mostly get away with this but changing tabs, for example, fails because it uses the state navigator (not the context one) and this is out of step