Created temporary navigation event
When fluent navigating from scene 1 to 3 then scene 2 doesn't have a navigation event until it's navigated to. Now that NavigationMotion listens to didShow instead of willShow, that navigation doesn't happen until later. So pressing back, or swiping back then scene 2 doesn't load until the swipe/back completes. This is too slow and shows a blank scene.
So, created a temporary navigation context if it's not already got one