Prevented blank push when double tap back
Press back twice fast and a blank scene was push in between processing the two backs. The first back sent message to js,;then the second back happens; the the js sends the first back back to native (onNavigate in js); this pushes a new scene because it's one crumb ahead now; there's no corresponding crumb on the server so the scene's blank; then js sends the second back back to native and pops the blank scene!?!
Flagged the js back navigation, in response to native back, as a history navigation. Then checked the history flag in js onNavigate listener and if it's true don't tell native about it - because native already knows!