Removed change tabs functionality
It's wasn't a good way to change tabs because navigated directly against the stateNavigators instead of using the context. It also caused problems when changing tabs straight away - seems that native fires to didShow's - one for the 0th crumb before the right one for the 1st crumb?!
First release won't have changing tabs. Think about it later and try to handle with context and maybe module instead