Moved navigate handlers out of component
It's too restrictive having the navigate handlers in the NavigationMotion component. For example, in a tabbed setup only the first tabs listeners are setup on load, it's not until visit another tab that the listeners are added. Can't guarantee tabs are loaded for deeplinking or notifications.
Also, can't guarantee that the first crumb is the first to load - might skip some - and then which crumb adds the listeners?