Calculated enter and exit anim in javascript
Imagine going from State A to State C fluently so that State B is in the middle. The exit State is A and the enter State is C, but A's next State is B. Similarly going back 2 from C to A, the exit State is C and enter is A, but A's next State is B.
This is different to Navigation React Mobile because render all Scenes in succession so next State is automatic. In Native, need to use the old crumbs, state and data to get the right next State and next data.