Got linking working on android
Created dedicated LinkActivity because using MainActivity doesn't work. Also abandoned using getInitialURL in javascript. Just used addEventListener('ur') and fired this event from the dedicated LinkActivity. Had to make sure LinkActivity launches the MainActivity for when the link comes in and the app isn't lauched. Passed the uri data to MainActivity so it can raise the url event