Got linking working on ios when app not open
Used getInitialURL (as intented). This didn't work on android because it meant using the MainActivity but not such problems on iOS. Call it outside of components - don't want any of this in components. Need to check it doesn't break android but think it's ok because 1. It's too early for android to return initial url (Activity not started yet) and 2. The LinkActivity never sets itself as currentActivity so getInitialURL should always return null
iOS and android behave a bit differently because android doesn't use getInitialURL - might mean they behave differently across Live Reload (iOS will run the initial navigation but android won't I think)