Remapped shared elements from old names
The shared elements might've changed since the scene was loaded. For example, zoom to 'green' details then change to 'blue' on details and go back to grid. Want to remap the shared element from green to blue. Used the old shared elements passed in (blue) to update the outgoing shared elements (was green). Then used the names (now blue) on the incoming to look up the blue shared element