Reverted to passing scene to children
The separate scene ReactDOM.render for each scene doesn't work for SSR. Need a single render to return all the HTML for SSR. Makes more sense for SceneProxy to be child of animated div instead of parent anyway. Also allows it to work with either multitple React roots or single. Good for migration and good to have choice