Displayed search bar children in controller
No idea what RCTShadowView is or why it works but copied it from RCTModalHostViewManager. SearchBar component works the same way as React Modal component - the child content is loaded into a separate UIViewController. Without the shadow view changes the main view displayed the child content and at a funny size.  The shadow changes the size but still need to style it absolute to prevent it taking up real estate (will do this in future commit)