Removed setState text infinite cycle
When type really fast the native and js can fight each other. The native sends first textChange 'r' and before the js responds send the next textChange 're'. Then the js responds from the first with 'r' and so native sends another textChange 'r' because it was 're'. But then js sends back the second textChange of 're' and so native sends textChange of 're' because it was 'r'. And so on!! Copied the locking mechanism from  RCTBaseTextInputView and TextInput where there's a count that the native sends to js and it only processes the setText if the counts match. When Fabric comes in this locking won't be needed because it'll be done sync