Cleared bar and controller when view removed
Happens when React component is unmounted. The tricky case is dev tools --> reload beause odd things happen to the navigation bar. Only odd case could find is reload on second screen and come back on the search bar isn't there but can navigate forward and back and it appears, or just do another reload and it comes back