Created NavigationStack component
Renders all the Scenes in the stack, one for each crumb plus the current state. Copied the getScenes from NavigationMotion in mobile because that's its counterpart. Will need the rest of the scene data when it comes to animating on android anyway - can add props to get the animation instead of adding to the state