Pushed subview onto navigationController
Can't call suerp insert because then the react native grumbles that the parent react view controller aren't right. It's expecting the parent to be the host view controller but it's actually the navigation controller. Still need to track the subviews because gonna need them when swipe back and navigate I think because need client and server subviews to match (can't call setState when going back)